+++ packages/example-floors-astro/src/components/SvgMapViewer/render.tsx
