+++ packages/example-floors-astro/src/components/SvgMapViewer/schema.ts
