+++ packages/example-floors-astro/src/components/SvgMapViewer/types.ts
