+++ packages/example-floors-astro/src/content.config.ts
@@ -2,7 +2,7 @@ import { file } from 'astro/loaders'
 import { defineCollection } from 'astro:content'
 
 import { collections as commonCollections } from '../node_modules/svgmapviewer-astro-floors/src/content.config'
-import { xinfoSchema } from './app/schema'
+import { xinfoSchema } from './components/SvgMapViewer/schema'
 
 const pois = defineCollection({
   loader: file('./src/content/pois.yaml'),
