+++ packages/example-floors-astro/src/pages/index.astro
@@ -1,10 +1,10 @@
 ---
 import { config } from '../utils/config'
 
-import { Main } from '../components/Main'
+import { SvgMapViewer } from '../components/SvgMapViewer/SvgMapViewer'
 import Layout from '../layouts/Layout.astro';
 ---
 
 <Layout title={config.title}>
-	<Main client:only="react" {...config} />
+	<SvgMapViewer client:only="react" {...config} />
 </Layout>
